created new module for blocks_affect in simulator
